refactor: extract BaseCommand
