+++ src/Commands/User.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace CodeIgniter\Shield\Commands;
 
-use CodeIgniter\CLI\BaseCommand;
 use CodeIgniter\Shield\Authentication\Authenticators\Session;
 use CodeIgniter\Shield\Commands\Exceptions\BadInputException;
 use CodeIgniter\Shield\Commands\Exceptions\CancelException;
-use CodeIgniter\Shield\Commands\Utils\InputOutput;
 use CodeIgniter\Shield\Config\Auth;
 use CodeIgniter\Shield\Entities\User as UserEntity;
 use CodeIgniter\Shield\Exceptions\UserNotFoundException;
@@ -18,19 +16,10 @@ use Config\Services;
 
 class User extends BaseCommand
 {
-    private static ?InputOutput $io = null;
-    private array $validActions     = [
+    private array $validActions = [
         'create', 'activate', 'deactivate', 'changename', 'changeemail',
         'delete', 'password', 'list', 'addgroup', 'removegroup',
     ];
-
-    /**
-     * The group the command is lumped under
-     * when listing commands.
-     *
-     * @var string
-     */
-    protected $group = 'Shield';
 
     /**
      * Command's name
@@ -249,31 +238,6 @@ class User extends BaseCommand
             'email'    => $rules['email'],
             'password' => $rules['password'],
         ];
-    }
-
-    /**
-     * Asks the user for input.
-     *
-     * @param string       $field      Output "field" question
-     * @param array|string $options    String to a default value, array to a list of options (the first option will be the default value)
-     * @param array|string $validation Validation rules
-     *
-     * @return string The user input
-     */
-    private function prompt(string $field, $options = null, $validation = null): string
-    {
-        return self::$io->prompt($field, $options, $validation);
-    }
-
-    /**
-     * Outputs a string to the cli on its own line.
-     */
-    private function write(
-        string $text = '',
-        ?string $foreground = null,
-        ?string $background = null
-    ): void {
-        self::$io->write($text, $foreground, $background);
     }
 
     /**
@@ -691,28 +655,5 @@ class User extends BaseCommand
         $this->checkUserExists($user);
 
         return $userModel->findById($user['id']);
-    }
-
-    private function ensureInputOutput(): void
-    {
-        if (self::$io === null) {
-            self::$io = new InputOutput();
-        }
-    }
-
-    /**
-     * @internal Testing purpose only
-     */
-    public static function setInputOutput(InputOutput $io): void
-    {
-        self::$io = $io;
-    }
-
-    /**
-     * @internal Testing purpose only
-     */
-    public static function resetInputOutput(): void
-    {
-        self::$io = null;
     }
 }
